clippy

+++ programs/margin/src/util.rs
@@ -326,7 +326,7 @@ mod test {
         assert_eq!(BitSet(0).max(), None);
         assert_eq!(BitSet(0).min(), None);
         for extremum in 0..8 {
-            let top = 2u8.checked_pow(extremum + 1 as u32).unwrap_or(u8::MAX);
+            let top = 2u8.checked_pow(extremum + 1).unwrap_or(u8::MAX);
             for byte in 2u8.pow(extremum as u32)..top {
                 assert_eq!(extremum as u32, BitSet(byte).max().unwrap());
                 assert_eq!(extremum as u32, BitSet(byte.reverse_bits()).min().unwrap());
